Revert "memberships: define & enforce permission policy and rules"
This reverts commit 13a8c772ea914223d193d64dfb0054e95438bf97.
