+++ invenio_communities/communities/__init__.py
@@ -2,7 +2,6 @@
 #
 # This file is part of Invenio.
 # Copyright (C) 2016-2021 CERN.
-# Copyright (C) 2022 Northwestern University.
 #
 # Invenio is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
@@ -10,10 +9,11 @@
 """Community Service API."""
 
 from .resources import CommunityResource, CommunityResourceConfig
-from .services import CommunityFileServiceConfig, CommunityService, \
-    CommunityServiceConfig, SearchOptions
+from .services import CommunityFileServiceConfig, CommunityNeed, \
+    CommunityService, CommunityServiceConfig, SearchOptions
 
 __all__ = (
+    'CommunityNeed',
     'CommunityService',
     'CommunityServiceConfig',
     'CommunityFileServiceConfig',
