+++ invenio_communities/communities/services/__init__.py
@@ -2,7 +2,6 @@
 #
 # This file is part of Invenio.
 # Copyright (C) 2016-2021 CERN.
-# Copyright (C) 2022 Northwestern University.
 #
 # Invenio is free software; you can redistribute it and/or modify it
 # under the terms of the MIT License; see LICENSE file for more details.
@@ -11,9 +10,11 @@
 
 from .config import CommunityFileServiceConfig, CommunityServiceConfig, \
     SearchOptions
+from .permissions import CommunityNeed
 from .service import CommunityService
 
 __all__ = (
+    'CommunityNeed',
     'CommunityService',
     'CommunityServiceConfig',
     'CommunityFileServiceConfig',
