+++ invenio_communities/members/records/mappings/v6/community_members/members-v1.0.0.json
@@ -1,0 +1,27 @@
+{
+  "mappings": {
+    "_doc": {
+      "numeric_detection": false,
+      "properties": {
+        "created": {
+          "type": "date"
+        },
+        "updated": {
+          "type": "date"
+        },
+        "id": {
+          "type": "keyword"
+        },
+        "community_id": {
+          "type": "keyword"
+        },
+        "user_id": {
+          "type": "keyword"
+        },
+        "role": {
+          "type": "keyword"
+        }
+      }
+    }
+  }
+}
