+++ invenio_communities/members/services/__init__.py
@@ -9,6 +9,7 @@
 
 
 from .config import MemberServiceConfig
+from .permissions import CommunityMembersPermissionPolicy
 from .schema import ROLE_TYPES
 from .service import MemberService
