+++ invenio_communities/members/services/config.py
@@ -9,9 +9,8 @@
 
 from invenio_records_resources.services import Link, RecordServiceConfig
 
-from ...communities.records.api import Community
-from ...permissions import CommunityPermissionPolicy
 from ..records import Member
+from .permissions import CommunityMembersPermissionPolicy
 from .schema import MemberSchema
 
 
@@ -32,11 +31,10 @@ class MemberLink(Link):
 class MemberServiceConfig(RecordServiceConfig):
     """Member Service Config."""
 
-    community_cls = Community
     record_cls = Member
     schema = MemberSchema
 
-    permission_policy_cls = CommunityPermissionPolicy
+    permission_policy_cls = CommunityMembersPermissionPolicy
     # result_item_cls = RequestEventItem
 
     # ResultItem configurations
